requirements.txt
matplotlib
seaborn
pandas
numpy
aiosqlite
discord.py
flask
python-dotenv
supabase
postgrest
